examples/requirements.txt
fastapi
uvicorn
requests
httpx
pillow
